% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finddxgap-package.R
\docType{package}
\name{finddxgap-package}
\alias{finddxgap}
\alias{finddxgap-package}
\title{finddxgap: What the Package Does (One Line, Title Case)}
\description{
To learn more about finddxgap, see the individual help files for each
function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikejohnpage.github.io/finddxgap/}
}

}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org}

Authors:
\itemize{
  \item Mike Page \email{mike@cynkra.com}
  \item Andrea Ranzato \email{andrea@cynkra.com}
}

Other contributors:
\itemize{
  \item FIND [copyright holder, funder]
}

}
\keyword{internal}
